refactor: Split up ensure.go in transformer
