+++ forst/internal/transformer/go/ensure.go
@@ -4,62 +4,10 @@ import (
 	"fmt"
 	"forst/internal/ast"
 	goast "go/ast"
-	"go/token"
 	"strings"
 
 	"github.com/sirupsen/logrus"
 )
-
-const (
-	// MinConstraint is the built-in Min constraint in Forst
-	MinConstraint = "Min"
-	// MaxConstraint is the built-in Max constraint in Forst
-	MaxConstraint = "Max"
-	// LessThanConstraint is the built-in LessThan constraint in Forst
-	LessThanConstraint = "LessThan"
-	// GreaterThanConstraint is the built-in GreaterThan constraint in Forst
-	GreaterThanConstraint = "GreaterThan"
-	// HasPrefixConstraint is the built-in HasPrefix constraint in Forst
-	HasPrefixConstraint = "HasPrefix"
-	// TrueConstraint is the built-in True constraint in Forst
-	TrueConstraint = "True"
-	// FalseConstraint is the built-in False constraint in Forst
-	FalseConstraint = "False"
-	// NilConstraint is the built-in Nil constraint in Forst
-	NilConstraint = "Nil"
-)
-
-const (
-	// BoolConstantTrue is the true constant in Go
-	BoolConstantTrue = "true"
-	// BoolConstantFalse is the false constant in Go
-	BoolConstantFalse = "false"
-	// NilConstant is the nil constant in Go
-	NilConstant = "nil"
-)
-
-// expectValue validates and returns a value node
-func (at *AssertionTransformer) expectValue(arg *ast.ConstraintArgumentNode) (ast.ValueNode, error) {
-	if arg == nil {
-		return nil, fmt.Errorf("expected an argument")
-	}
-
-	if arg.Value == nil {
-		return nil, fmt.Errorf("expected argument to be a value")
-	}
-
-	return *arg.Value, nil
-}
-
-// AssertionTransformer handles the transformation of assertions
-type AssertionTransformer struct {
-	transformer *Transformer
-}
-
-// NewAssertionTransformer creates a new AssertionTransformer
-func NewAssertionTransformer(t *Transformer) *AssertionTransformer {
-	return &AssertionTransformer{transformer: t}
-}
 
 // transformEnsureConstraints transforms an ensure node based on its type
 func (at *AssertionTransformer) transformEnsureConstraints(ensure ast.EnsureNode) (goast.Expr, error) {
@@ -89,299 +37,6 @@ func (at *AssertionTransformer) transformEnsureConstraints(ensure ast.EnsureNode
 	}
 }
 
-// validateConstraintArgs validates the number of arguments for a constraint
-func (at *AssertionTransformer) validateConstraintArgs(constraint ast.ConstraintNode, expectedArgs int) error {
-	if len(constraint.Args) != expectedArgs {
-		return fmt.Errorf("%s constraint requires %d argument(s)", constraint.Name, expectedArgs)
-	}
-	return nil
-}
-
-// transformStringEnsure transforms a string ensure
-func (at *AssertionTransformer) transformStringEnsure(ensure ast.EnsureNode) (goast.Expr, error) {
-	result := []goast.Expr{}
-
-	for _, constraint := range ensure.Assertion.Constraints {
-		var expr goast.Expr
-
-		switch constraint.Name {
-		case MinConstraint:
-			if err := at.validateConstraintArgs(constraint, 1); err != nil {
-				return nil, err
-			}
-			arg, err := at.expectValue(&constraint.Args[0])
-			if err != nil {
-				return nil, err
-			}
-			arg, err = expectIntLiteral(arg)
-			if err != nil {
-				return nil, err
-			}
-			expr = &goast.BinaryExpr{
-				X: &goast.CallExpr{
-					Fun: goast.NewIdent("len"),
-					Args: []goast.Expr{
-						at.transformer.transformExpression(ensure.Variable),
-					},
-				},
-				Op: token.LSS,
-				Y:  at.transformer.transformExpression(arg),
-			}
-		case MaxConstraint:
-			if err := at.validateConstraintArgs(constraint, 1); err != nil {
-				return nil, err
-			}
-			arg, err := at.expectValue(&constraint.Args[0])
-			if err != nil {
-				return nil, err
-			}
-			arg, err = expectIntLiteral(arg)
-			if err != nil {
-				return nil, err
-			}
-			expr = &goast.BinaryExpr{
-				X: &goast.CallExpr{
-					Fun: goast.NewIdent("len"),
-					Args: []goast.Expr{
-						at.transformer.transformExpression(ensure.Variable),
-					},
-				},
-				Op: token.GTR,
-				Y:  at.transformer.transformExpression(arg),
-			}
-		case HasPrefixConstraint:
-			if err := at.validateConstraintArgs(constraint, 1); err != nil {
-				return nil, err
-			}
-			arg, err := at.expectValue(&constraint.Args[0])
-			if err != nil {
-				return nil, err
-			}
-			arg, err = expectStringLiteral(arg)
-			if err != nil {
-				return nil, err
-			}
-			expr = negateCondition(&goast.CallExpr{
-				Fun: &goast.SelectorExpr{
-					X:   goast.NewIdent("strings"),
-					Sel: goast.NewIdent("HasPrefix"),
-				},
-				Args: []goast.Expr{
-					at.transformer.transformExpression(ensure.Variable),
-					at.transformer.transformExpression(arg),
-				},
-			})
-		default:
-			return nil, fmt.Errorf("unknown String constraint: %s", constraint.Name)
-		}
-
-		result = append(result, expr)
-	}
-	return disjoin(result), nil
-}
-
-// transformIntEnsure transforms an integer assertion
-func (at *AssertionTransformer) transformIntEnsure(ensure ast.EnsureNode) (goast.Expr, error) {
-	result := []goast.Expr{}
-	for _, constraint := range ensure.Assertion.Constraints {
-		var expr goast.Expr
-
-		switch constraint.Name {
-		case MinConstraint:
-			if err := at.validateConstraintArgs(constraint, 1); err != nil {
-				return nil, err
-			}
-			arg, err := at.expectValue(&constraint.Args[0])
-			if err != nil {
-				return nil, err
-			}
-			arg, err = expectNumberLiteral(arg)
-			if err != nil {
-				return nil, err
-			}
-			expr = &goast.BinaryExpr{
-				X:  at.transformer.transformExpression(ensure.Variable),
-				Op: token.LSS,
-				Y:  at.transformer.transformExpression(arg),
-			}
-		case MaxConstraint:
-			if err := at.validateConstraintArgs(constraint, 1); err != nil {
-				return nil, err
-			}
-			arg, err := at.expectValue(&constraint.Args[0])
-			if err != nil {
-				return nil, err
-			}
-			arg, err = expectNumberLiteral(arg)
-			if err != nil {
-				return nil, err
-			}
-			expr = &goast.BinaryExpr{
-				X:  at.transformer.transformExpression(ensure.Variable),
-				Op: token.GTR,
-				Y:  at.transformer.transformExpression(arg),
-			}
-		case LessThanConstraint:
-			if err := at.validateConstraintArgs(constraint, 1); err != nil {
-				return nil, err
-			}
-			arg, err := at.expectValue(&constraint.Args[0])
-			if err != nil {
-				return nil, err
-			}
-			arg, err = expectNumberLiteral(arg)
-			if err != nil {
-				return nil, err
-			}
-			expr = &goast.BinaryExpr{
-				X:  at.transformer.transformExpression(ensure.Variable),
-				Op: token.GEQ,
-				Y:  at.transformer.transformExpression(arg),
-			}
-		case GreaterThanConstraint:
-			if err := at.validateConstraintArgs(constraint, 1); err != nil {
-				return nil, err
-			}
-			arg, err := at.expectValue(&constraint.Args[0])
-			if err != nil {
-				return nil, err
-			}
-			arg, err = expectNumberLiteral(arg)
-			if err != nil {
-				return nil, err
-			}
-			expr = &goast.BinaryExpr{
-				X:  at.transformer.transformExpression(ensure.Variable),
-				Op: token.LEQ,
-				Y:  at.transformer.transformExpression(arg),
-			}
-		default:
-			return nil, fmt.Errorf("unknown Int constraint: %s", constraint.Name)
-		}
-
-		result = append(result, expr)
-	}
-	return disjoin(result), nil
-}
-
-// transformFloatEnsure transforms a float assertion
-func (at *AssertionTransformer) transformFloatEnsure(ensure ast.EnsureNode) (goast.Expr, error) {
-	result := []goast.Expr{}
-	for _, constraint := range ensure.Assertion.Constraints {
-		var expr goast.Expr
-
-		switch constraint.Name {
-		case MinConstraint:
-			if err := at.validateConstraintArgs(constraint, 1); err != nil {
-				return nil, err
-			}
-			arg, err := at.expectValue(&constraint.Args[0])
-			if err != nil {
-				return nil, err
-			}
-			arg, err = expectNumberLiteral(arg)
-			if err != nil {
-				return nil, err
-			}
-			expr = &goast.BinaryExpr{
-				X:  at.transformer.transformExpression(ensure.Variable),
-				Op: token.LSS,
-				Y:  at.transformer.transformExpression(arg),
-			}
-		case MaxConstraint:
-			if err := at.validateConstraintArgs(constraint, 1); err != nil {
-				return nil, err
-			}
-			arg, err := at.expectValue(&constraint.Args[0])
-			if err != nil {
-				return nil, err
-			}
-			arg, err = expectNumberLiteral(arg)
-			if err != nil {
-				return nil, err
-			}
-			expr = &goast.BinaryExpr{
-				X:  at.transformer.transformExpression(ensure.Variable),
-				Op: token.GTR,
-				Y:  at.transformer.transformExpression(arg),
-			}
-		case GreaterThanConstraint:
-			if err := at.validateConstraintArgs(constraint, 1); err != nil {
-				return nil, err
-			}
-			arg, err := at.expectValue(&constraint.Args[0])
-			if err != nil {
-				return nil, err
-			}
-			arg, err = expectNumberLiteral(arg)
-			if err != nil {
-				return nil, err
-			}
-			expr = &goast.BinaryExpr{
-				X:  at.transformer.transformExpression(ensure.Variable),
-				Op: token.LEQ,
-				Y:  at.transformer.transformExpression(arg),
-			}
-		default:
-			return nil, fmt.Errorf("unknown Float constraint: %s", constraint.Name)
-		}
-
-		result = append(result, expr)
-	}
-	return disjoin(result), nil
-}
-
-// transformBoolEnsure transforms a boolean assertion
-func (at *AssertionTransformer) transformBoolEnsure(ensure ast.EnsureNode) (goast.Expr, error) {
-	result := []goast.Expr{}
-	for _, constraint := range ensure.Assertion.Constraints {
-		var expr goast.Expr
-
-		switch constraint.Name {
-		case TrueConstraint:
-			if err := at.validateConstraintArgs(constraint, 0); err != nil {
-				return nil, err
-			}
-			expr = negateCondition(at.transformer.transformExpression(ensure.Variable))
-		case FalseConstraint:
-			if err := at.validateConstraintArgs(constraint, 0); err != nil {
-				return nil, err
-			}
-			expr = at.transformer.transformExpression(ensure.Variable)
-		default:
-			return nil, fmt.Errorf("unknown Bool constraint: %s", constraint.Name)
-		}
-
-		result = append(result, expr)
-	}
-	return disjoin(result), nil
-}
-
-// transformErrorEnsure transforms an error assertion
-func (at *AssertionTransformer) transformErrorEnsure(ensure ast.EnsureNode) (goast.Expr, error) {
-	result := []goast.Expr{}
-	for _, constraint := range ensure.Assertion.Constraints {
-		var expr goast.Expr
-
-		switch constraint.Name {
-		case NilConstraint:
-			if err := at.validateConstraintArgs(constraint, 0); err != nil {
-				return nil, err
-			}
-			expr = &goast.BinaryExpr{
-				X:  at.transformer.transformExpression(ensure.Variable),
-				Op: token.NEQ,
-				Y:  goast.NewIdent(NilConstant),
-			}
-		default:
-			return nil, fmt.Errorf("unknown Error constraint: %s", constraint.Name)
-		}
-
-		result = append(result, expr)
-	}
-	return disjoin(result), nil
-}
-
 func (t *Transformer) getEnsureBaseType(ensure ast.EnsureNode) (ast.TypeNode, error) {
 	if ensure.Assertion.BaseType != nil {
 		return ast.TypeNode{Ident: *ensure.Assertion.BaseType}, nil
@@ -392,30 +47,6 @@ func (t *Transformer) getEnsureBaseType(ensure ast.EnsureNode) (ast.TypeNode, er
 		return ast.TypeNode{}, err
 	}
 	return *ensureBaseType, nil
-}
-
-// Helper to look up a TypeGuardNode by name
-func (t *Transformer) lookupTypeGuardNode(name string) *ast.TypeGuardNode {
-	for _, def := range t.TypeChecker.Defs {
-		if tg, ok := def.(ast.TypeGuardNode); ok {
-			t.log.WithFields(logrus.Fields{
-				"requested": name,
-				"candidate": tg.GetIdent(),
-			}).Trace("lookupTypeGuardNode: candidate check")
-			if tg.GetIdent() == name {
-				t.log.WithFields(logrus.Fields{
-					"requested": name,
-					"found":     true,
-				}).Trace("lookupTypeGuardNode: found match")
-				return &tg
-			}
-		}
-	}
-	t.log.WithFields(logrus.Fields{
-		"requested": name,
-		"found":     false,
-	}).Trace("lookupTypeGuardNode: not found")
-	panic("Type guard not found: " + name)
 }
 
 func (t *Transformer) transformEnsureCondition(ensure ast.EnsureNode) (goast.Expr, error) {
@@ -499,39 +130,6 @@ func (t *Transformer) transformEnsureCondition(ensure ast.EnsureNode) (goast.Exp
 	return expr, nil
 }
 
-func (t *Transformer) transformTypeGuardEnsure(ensure ast.EnsureNode) goast.Expr {
-	// Look up the real type guard node by name
-	guardName := ensure.Assertion.Constraints[0].Name
-	typeGuardNode := t.lookupTypeGuardNode(guardName)
-
-	// Use hash-based guard function name
-	hash := t.TypeChecker.Hasher.HashNode(*typeGuardNode)
-	guardFuncName := hash.ToGuardIdent()
-
-	if len(typeGuardNode.Parameters()) > 0 {
-		switch typeGuardNode.Parameters()[0].(type) {
-		case ast.SimpleParamNode:
-			return &goast.CallExpr{
-				Fun: goast.NewIdent(string(guardFuncName)),
-				Args: []goast.Expr{
-					t.transformExpression(ensure.Variable),
-				},
-			}
-		case ast.DestructuredParamNode:
-			panic("DestructuredParamNode not supported in type guard assertion")
-		}
-	} else {
-		panic("Type guard has no parameters")
-	}
-
-	return &goast.CallExpr{
-		Fun: goast.NewIdent(string(guardFuncName)),
-		Args: []goast.Expr{
-			t.transformExpression(ensure.Variable),
-		},
-	}
-}
-
 func (t *AssertionTransformer) transformEnsureCondition(ensure *ast.EnsureNode) ([]goast.Stmt, error) {
 	// Get the variable type from the symbol table
 	varType, err := t.transformer.TypeChecker.LookupVariableType(&ensure.Variable, t.transformer.currentScope())
@@ -602,43 +200,4 @@ func (t *AssertionTransformer) transformEnsureCondition(ensure *ast.EnsureNode) 
 	}
 
 	return transformedStmts, nil
-}
-
-func (t *AssertionTransformer) isTypeGuardCompatible(varType ast.TypeNode, typeGuard *ast.TypeGuardNode) bool {
-	t.transformer.log.WithFields(logrus.Fields{
-		"varType":   varType.Ident,
-		"typeGuard": typeGuard.GetIdent(),
-	}).Trace("[isTypeGuardCompatible] Checking type guard compatibility")
-
-	// Get the base type of the variable
-	baseType, err := t.transformer.getEnsureBaseType(ast.EnsureNode{Variable: ast.VariableNode{ExplicitType: varType}})
-	if err != nil {
-		t.transformer.log.WithError(err).Error("[isTypeGuardCompatible] Failed to get base type")
-		return false
-	}
-	t.transformer.log.WithFields(logrus.Fields{
-		"baseType": baseType.Ident,
-	}).Trace("[isTypeGuardCompatible] Base type")
-
-	// Check if the type guard is defined for the base type
-	for _, param := range typeGuard.Parameters() {
-		t.transformer.log.WithFields(logrus.Fields{
-			"paramType": param.GetType().Ident,
-			"baseType":  baseType.Ident,
-		}).Trace("[isTypeGuardCompatible] Checking parameter type")
-
-		if param.GetType().Ident == baseType.Ident {
-			t.transformer.log.WithFields(logrus.Fields{
-				"typeGuard": typeGuard.GetIdent(),
-				"baseType":  baseType.Ident,
-			}).Trace("[isTypeGuardCompatible] Found compatible type guard")
-			return true
-		}
-	}
-
-	t.transformer.log.WithFields(logrus.Fields{
-		"typeGuard": typeGuard.GetIdent(),
-		"baseType":  baseType.Ident,
-	}).Trace("[isTypeGuardCompatible] No compatible type guard found")
-	return false
 }
